+++ packages/component/src/control/logo.ts
@@ -1,6 +1,7 @@
 import { IControlOption } from '@antv/l7-core';
 import { DOM } from '@antv/l7-utils';
 import Control, { PositionType } from './BaseControl';
+
 export default class Logo extends Control {
   public getDefault() {
     return {
